+++ src/openassetio-python/tests/bridge/python/test_converter.cpp
@@ -262,8 +262,7 @@ using CastableClasses = std::tuple<
     log::SeverityFilter,
     managerApi::Host,
     managerApi::HostSession,
-    managerApi::ManagerInterface,
-    managerApi::ManagerStateBase
+    managerApi::ManagerInterface
 >;
 // clang-format on
 }  // namespace